+++ tinper-bee/bee-colorpicker/CHANGELOG.md
@@ -1,3 +1,8 @@
+<a name="0.0.10"></a>
+## [0.0.10](https://github.com/tinper-bee/bee-colorpicker/compare/v0.0.9...v0.0.10) (2019-05-28)
+
+
+
 <a name="0.0.9"></a>
 ## [0.0.9](https://github.com/tinper-bee/bee-colorpicker/compare/v0.0.8...v0.0.9) (2019-04-28)